+++ packages/services/src/state_committer.rs
@@ -21,7 +21,7 @@ pub struct Config {
     pub tx_max_fee: u128,
 }
 
-#[cfg(test)]
+#[cfg(feature = "test-helpers")]
 impl Default for Config {
     fn default() -> Self {
         Self {
@@ -242,294 +242,6 @@ where
             self.submit_fragments_if_ready().await?
         };
 
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::time::Duration;
-
-    use clock::{SystemClock, TestClock};
-
-    use super::*;
-    use crate::{test_utils, Runner, StateCommitter};
-
-    #[tokio::test]
-    async fn submits_fragments_when_required_count_accumulated() -> Result<()> {
-        // given
-        let setup = test_utils::Setup::init().await;
-
-        let fragments = setup.insert_fragments(0, 4).await;
-
-        let tx_hash = [0; 32];
-        let l1_mock_submit = test_utils::mocks::l1::expects_state_submissions([(
-            Some(NonEmpty::from_vec(fragments.clone()).unwrap()),
-            tx_hash,
-        )]);
-
-        let fuel_mock = test_utils::mocks::fuel::latest_height_is(0);
-        let mut state_committer = StateCommitter::new(
-            l1_mock_submit,
-            fuel_mock,
-            setup.db(),
-            Config {
-                lookback_window: 1000,
-                fragment_accumulation_timeout: Duration::from_secs(60),
-                fragments_to_accumulate: 4.try_into().unwrap(),
-                ..Default::default()
-            },
-            TestClock::default(),
-        );
-
-        // when
-        state_committer.run().await?;
-
-        // then
-        // Mocks validate that the fragments have been sent
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn submits_fragments_on_timeout_before_accumulation() -> Result<()> {
-        // given
-        let clock = TestClock::default();
-        let setup = test_utils::Setup::init().await;
-
-        let fragments = setup.insert_fragments(0, 5).await; // Only 5 fragments, less than required
-
-        let tx_hash = [1; 32];
-        let l1_mock_submit = test_utils::mocks::l1::expects_state_submissions([(
-            Some(NonEmpty::from_vec(fragments.clone()).unwrap()),
-            tx_hash,
-        )]);
-
-        let fuel_mock = test_utils::mocks::fuel::latest_height_is(0);
-        let mut state_committer = StateCommitter::new(
-            l1_mock_submit,
-            fuel_mock,
-            setup.db(),
-            Config {
-                lookback_window: 1000,
-                fragment_accumulation_timeout: Duration::from_secs(60),
-                fragments_to_accumulate: 10.try_into().unwrap(),
-                ..Default::default()
-            },
-            clock.clone(),
-        );
-
-        // Advance time beyond the timeout
-        clock.advance_time(Duration::from_secs(61));
-
-        // when
-        state_committer.run().await?;
-
-        // then
-        // Mocks validate that the fragments have been sent despite insufficient accumulation
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn does_not_submit_fragments_before_required_count_or_timeout() -> Result<()> {
-        // given
-        let clock = TestClock::default();
-        let setup = test_utils::Setup::init().await;
-
-        let _fragments = setup.insert_fragments(0, 5).await; // Only 5 fragments, less than required
-
-        let l1_mock_submit = test_utils::mocks::l1::expects_state_submissions([]); // Expect no submissions
-
-        let fuel_mock = test_utils::mocks::fuel::latest_height_is(0);
-        let mut state_committer = StateCommitter::new(
-            l1_mock_submit,
-            fuel_mock,
-            setup.db(),
-            Config {
-                lookback_window: 1000,
-                fragment_accumulation_timeout: Duration::from_secs(60),
-                fragments_to_accumulate: 10.try_into().unwrap(),
-                ..Default::default()
-            },
-            clock.clone(),
-        );
-
-        // Advance time less than the timeout
-        clock.advance_time(Duration::from_secs(30));
-
-        // when
-        state_committer.run().await?;
-
-        // then
-        // Mocks validate that no fragments have been sent
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn submits_fragments_when_required_count_before_timeout() -> Result<()> {
-        // given
-        let clock = TestClock::default();
-        let setup = test_utils::Setup::init().await;
-
-        let fragments = setup.insert_fragments(0, 5).await;
-
-        let tx_hash = [3; 32];
-        let l1_mock_submit = test_utils::mocks::l1::expects_state_submissions([(
-            Some(NonEmpty::from_vec(fragments).unwrap()),
-            tx_hash,
-        )]);
-
-        let fuel_mock = test_utils::mocks::fuel::latest_height_is(0);
-        let mut state_committer = StateCommitter::new(
-            l1_mock_submit,
-            fuel_mock,
-            setup.db(),
-            Config {
-                lookback_window: 1000,
-                fragment_accumulation_timeout: Duration::from_secs(60),
-                fragments_to_accumulate: 5.try_into().unwrap(),
-                ..Default::default()
-            },
-            clock.clone(),
-        );
-
-        // when
-        state_committer.run().await?;
-
-        // then
-        // Mocks validate that the fragments have been sent
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn timeout_measured_from_last_finalized_fragment() -> Result<()> {
-        // given
-        let clock = TestClock::default();
-        let setup = test_utils::Setup::init().await;
-
-        // Insert initial fragments
-        setup.commit_single_block_bundle(clock.now()).await;
-
-        let fragments_to_submit = setup.insert_fragments(1, 2).await;
-
-        let tx_hash = [4; 32];
-        let l1_mock_submit = test_utils::mocks::l1::expects_state_submissions([(
-            Some(NonEmpty::from_vec(fragments_to_submit).unwrap()),
-            tx_hash,
-        )]);
-
-        let fuel_mock = test_utils::mocks::fuel::latest_height_is(1);
-        let mut state_committer = StateCommitter::new(
-            l1_mock_submit,
-            fuel_mock,
-            setup.db(),
-            Config {
-                lookback_window: 1000,
-                fragment_accumulation_timeout: Duration::from_secs(60),
-                fragments_to_accumulate: 10.try_into().unwrap(),
-                ..Default::default()
-            },
-            clock.clone(),
-        );
-
-        // Advance time to exceed the timeout since last finalized fragment
-        clock.advance_time(Duration::from_secs(60));
-
-        // when
-        state_committer.run().await?;
-
-        // then
-        // Mocks validate that the fragments were sent even though the accumulation target was not reached
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn timeout_measured_from_startup_if_no_finalized_fragment() -> Result<()> {
-        // given
-        let clock = TestClock::default();
-        let setup = test_utils::Setup::init().await;
-
-        let fragments = setup.insert_fragments(0, 5).await; // Only 5 fragments, less than required
-
-        let tx_hash = [5; 32];
-        let l1_mock_submit = test_utils::mocks::l1::expects_state_submissions([(
-            Some(NonEmpty::from_vec(fragments.clone()).unwrap()),
-            tx_hash,
-        )]);
-
-        let fuel_mock = test_utils::mocks::fuel::latest_height_is(0);
-        let mut state_committer = StateCommitter::new(
-            l1_mock_submit,
-            fuel_mock,
-            setup.db(),
-            Config {
-                lookback_window: 1000,
-                fragment_accumulation_timeout: Duration::from_secs(60),
-                fragments_to_accumulate: 10.try_into().unwrap(),
-                ..Default::default()
-            },
-            clock.clone(),
-        );
-
-        // Advance time beyond the timeout from startup
-        clock.advance_time(Duration::from_secs(61));
-
-        // when
-        state_committer.run().await?;
-
-        // then
-        // Mocks validate that the fragments have been sent despite insufficient accumulation
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn resubmits_fragments_when_gas_bump_timeout_exceeded() -> Result<()> {
-        // given
-        // init test clock with actual time to be able to compare with db timestamps
-        let clock = TestClock::new(SystemClock {}.now());
-        let setup = test_utils::Setup::init().await;
-
-        let fragments = setup.insert_fragments(0, 5).await;
-
-        let tx_hash_1 = [6; 32];
-        let tx_hash_2 = [7; 32];
-        let l1_mock_submit = test_utils::mocks::l1::expects_state_submissions([
-            (
-                Some(NonEmpty::from_vec(fragments.clone()).unwrap()),
-                tx_hash_1,
-            ),
-            (
-                Some(NonEmpty::from_vec(fragments.clone()).unwrap()),
-                tx_hash_2,
-            ),
-        ]);
-
-        let fuel_mock = test_utils::mocks::fuel::latest_height_is(0);
-        let mut state_committer = StateCommitter::new(
-            l1_mock_submit,
-            fuel_mock,
-            setup.db(),
-            Config {
-                lookback_window: 1000,
-                fragment_accumulation_timeout: Duration::from_secs(60),
-                fragments_to_accumulate: 5.try_into().unwrap(),
-                gas_bump_timeout: Duration::from_secs(60),
-                ..Default::default()
-            },
-            clock.clone(),
-        );
-
-        // Submit the initial fragments
-        state_committer.run().await?;
-
-        // Advance time beyond the gas bump timeout, we need to advance a bit more
-        // because the db clock record the transaction
-        clock.advance_time(Duration::from_secs(80));
-
-        // when
-        state_committer.run().await?;
-
-        // then
-        // Mocks validate that the fragments have been sent again
         Ok(())
     }
 }
